--- PassOn/Engine/Extensions/CopyExtensions.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Linq;
-using System.Reflection.Emit;
-using System.Reflection;
-using System.Runtime.Serialization;
-using System.Text;
-using System.Threading.Tasks;
-using PassOn.Utilities;
-
-namespace PassOn.EngineExtensions
-{
-    internal static class CopyExtensions
-    {       
-
-        
-    }
-}
